requirements.txt
# F1Tenth Reinforcement Learning Project Dependencies

# Version tools for older gym compatibility
setuptools==65.5.0
wheel==0.38.4

# Core ML/RL Libraries
# Note: gym 0.21.0 or higher needs older setuptools to build correctly
gym==0.23.1
stable_baselines3[extra]
shimmy
pyyaml

# Visualization & Media
imageio
imageio-ffmpeg
matplotlib

# Deep Learning
# numpy<=1.22.0 is preferred for f110_gym but compatible versions can vary
numpy
torch
